# docs/phase2/openapi-v2.yaml
openapi: 3.0.3
info:
  title: TeamCollab MVP API
  version: 2.0.0
  description: >
    团队协作工具 MVP REST API（看板/任务/成员/评论/通知）。Phase 2 契约，前后端联调唯一依据。

    统一响应格式：{ "code": 0, "data": {}, "message": "" }，code=0 成功。
    错误码：40001 参数校验失败 / 40101 未认证或 token 过期 / 40301 无权限(RBAC) /
    40401 资源不存在(含跨租户访问，不泄露存在性) / 40901 资源冲突 / 42901 频率限制 / 50000 服务器内部错误。

    认证：Authorization: Bearer <access_token>。SSE 端点因 EventSource 无法自定义 Header，走 query token。

    版本锚定：next@15.5.9+ / fastapi@0.140.x / lucide-react@1.24.0（全项目唯一图标库，禁止 emoji 图标）。
servers:
  - url: http://localhost:8000
    description: FastAPI 本地开发
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: 注册（创建用户 + 默认团队）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功，返回用户信息 + token 对
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          description: 邮箱已注册（code=40901），不泄露账户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 登录（OAuth2 Password Flow，application/x-www-form-urlencoded）
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string, format: email, description: 邮箱 }
                password: { type: string }
      responses:
        '200':
          description: 登录成功，返回 access + refresh
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: 凭证错误（code=40101）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/me:
    get:
      tags: [auth]
      summary: 当前用户信息 + 所属团队 + 角色
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: 未认证（code=40101）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/me/tasks:
    get:
      tags: [tasks]
      summary: 我的任务聚合（跨项目，只返回分配给当前用户的任务）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string, enum: [todo, in_progress, done] }, description: 按状态筛选 }
        - { name: q, in: query, required: false, schema: { type: string }, description: 关键词（标题/描述 ILIKE） }
        - { name: page, in: query, required: false, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: 分页任务列表（逾期置顶由前端按 due_date 处理）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasks'
        '401':
          description: 未认证（code=40101）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ---------------- TEAMS ----------------
...
  /api/v1/teams/{teamId}/boards:
    get:
      tags: [boards]
      summary: 看板列表（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: 看板列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
        '403':
          description: 无权限（code=40301）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [boards]
      summary: 创建看板（member+）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 100 }
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
        '400':
          description: 参数校验失败（code=40001）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/notifications:
    get:
      tags: [notifications]
      summary: 历史通知列表（登录用户自己的，按 created_at 倒序）
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, required: false, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: unread, in: query, required: false, schema: { type: boolean }, description: 仅未读（true 时过滤 is_read=false） }
      responses:
        '200':
          description: 分页通知列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedNotifications'
        '401':
          description: 未认证（code=40101）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, data, message]
      properties:
        code:
          type: integer
          enum: [40001, 40101, 40301, 40401, 40901, 42901, 50000]
          description: 非 0 即错误
        data:
          type: object
          nullable: true
          description: 错误时恒为 null
        message:
          type: string
          description: 人类可读错误描述（不泄露内部细节）
...
    Role:
      type: string
      enum: [owner, admin, member, viewer]
      description: owner=团队创建者(不可降级) / admin=管理员 / member=成员 / viewer=只读
    Priority:
      type: string
      enum: [low, medium, high, urgent]
    TaskStatus:
      type: string
      enum: [todo, in_progress, done]
    NotificationType:
      type: string
      enum: [task_assigned, task_moved, comment_added, task_due_soon]
      description: task_assigned=被分配 / task_moved=任务被移动 / comment_added=被@提及或评论 / task_due_soon=距截止≤24h

    # ---------------- 认证 ----------------
    RegisterRequest:
      type: object
      required: [email, password, display_name]
      properties:
        email: { type: string, format: email, maxLength: 255 }
        password: { type: string, minLength: 8, maxLength: 72, description: bcrypt 输入上限 72 字节 }
        display_name: { type: string, minLength: 1, maxLength: 100 }
    AuthResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            user: { $ref: '#/components/schemas/User' }
            team: { $ref: '#/components/schemas/Team' }
            role: { $ref: '#/components/schemas/Role' }
            access_token: { type: string }
            refresh_token: { type: string }
            token_type: { type: string, example: bearer }
            expires_in: { type: integer, example: 900, description: access token 有效期（秒），MVP 15 分钟 }
        message: { type: string, example: '' }
...
    MeResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            user: { $ref: '#/components/schemas/User' }
            team: { $ref: '#/components/schemas/Team' }
            role: { $ref: '#/components/schemas/Role' }
        message: { type: string, example: '' }

    # ---------------- 用户/团队/成员 ----------------
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        created_at: { type: string, format: date-time }
    Team:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
...
    Board:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        name: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    BoardResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data: { $ref: '#/components/schemas/Board' }
        message: { type: string, example: '' }
    BoardListResponse:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: array
          items: { $ref: '#/components/schemas/Board' }
        message: { type: string, example: '' }
...
    Task:
      type: object
      properties:
        id: { type: string, format: uuid }
        board_id: { type: string, format: uuid }
        board_name: { type: string, description: 我的任务聚合时附带，便于跳转 }
        column_id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string, nullable: true }
        assignee_id: { type: string, format: uuid, nullable: true }
        assignee:
          type: object
          nullable: true
          properties:
            id: { type: string, format: uuid }
            display_name: { type: string }
        priority: { $ref: '#/components/schemas/Priority' }
        status: { $ref: '#/components/schemas/TaskStatus' }
        position: { type: integer }
        due_date: { type: string, format: date, nullable: true }
        start_date: { type: string, format: date, nullable: true }
        created_by: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PaginatedTasks:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Task' }
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            hasMore: { type: boolean }
        message: { type: string, example: '' }

    # ---------------- 评论 ----------------
...
    Notification:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/NotificationType' }
        payload:
          type: object
          description: >
            按 type 约定：task_assigned 含 { task_id, task_title, board_id, actor_id }；
            task_moved 含 { task_id, task_title, board_id, from_column_id, to_column_id, actor_id }；
            comment_added 含 { task_id, task_title, comment_id, comment_snippet, actor_id }；
            task_due_soon 含 { task_id, task_title, due_date }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }
...
    PaginatedNotifications:
      type: object
      required: [code, data, message]
      properties:
        code: { type: integer, example: 0 }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/Notification' }
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
            hasMore: { type: boolean }
        message: { type: string, example: '' }